Reland: Fix bug making us not inline when a simple class without clinit had static fields with nothing setting them
The conditional was reversed, so if there was fields, we would only
allow inlining if there was a field that had an allocation side effect.

Fix MethodHandle inlining. We simply newer inline methods that has const method handles in them (javac will not produce these)

This gives 50K on chrome

Bug: 122683370, 123358657
Change-Id: Iecb187b54cee4930eb4a96ca4fc57fefc8bfd285